retour en arriere

+++ philo_one/srcs/print.c
@@ -12,10 +12,12 @@
 
 #include "philo_one.h"
 
+
+
 void	print_state(t_philo *philo, int is_dead)
 {
 	char	**outputs;
-	int		ts;
+	long	ts;
 
 	outputs = (char *[TOTAL_STATES]){"is eating", "is sleeping",
 		"is thinking", "has taken a fork", "died"};
@@ -23,9 +25,9 @@ void	print_state(t_philo *philo, int is_dead)
 	{
 		ts = get_timestamp(philo->start_ts);
 		if (philo->state != EAT)
-			printf("%d %d %s\n", ts, philo->num, outputs[philo->state]);
+			printf("%ld %d %s\n", ts, philo->num, outputs[philo->state]);
 		else
-			printf("%d %d %s\n%d %d %s\n%d %d %s\n", ts,
+			printf("%ld %d %s\n%ld %d %s\n%ld %d %s\n", ts,
 				philo->num, outputs[HAS_FORKS], ts,
 				philo->num, outputs[HAS_FORKS], ts, philo->num, outputs[EAT]);
 		if (!is_dead)
